Refine video-recorder.js and put associate video handler to homeWin
